Add architecture clarification statement. (#567)
* Add architecture clarification statement.

Adding a paragraph to explain the intent the group has in providing the given component architecture, as well as the expectations for implementers. This was originally https://github.com/w3c-ccg/vcalm/pull/562 but due to bad git practices on my part I was forced to reset. 

The current statement should include the discussions and suggestions from PR562.

* Editorial changes

Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>

---------

Co-authored-by: Patrick St-Louis <43082425+PatStLouis@users.noreply.github.com>
Co-authored-by: Ted Thibodeau Jr <tthibodeau@openlinksw.com>